Create MVC Pattern for Controllers and Routes
