+++ handlers/weatherHandler.js
@@ -1,0 +1,49 @@
+function getRandomWeather(min, max) {
+  return Math.floor(Math.random() * (max - min) + min);
+}
+function getRandomForecast() {
+  let forecast = [
+    "Freezing",
+    "Bracing",
+    "Chilly",
+    "Cool",
+    "Mild",
+    "Warm",
+    "Balmy",
+    "Hot",
+    "Sweltering",
+    "Scorching",
+  ];
+  let random_val = Math.floor(Math.random() * forecast.length);
+  return forecast[random_val];
+}
+function getDatetime() {
+  let currentdate = new Date();
+  let datetime =
+    "Last Sync: " +
+    currentdate.getDate() +
+    "/" +
+    (currentdate.getMonth() + 1) +
+    "/" +
+    currentdate.getFullYear() +
+    " @ " +
+    currentdate.getHours() +
+    ":" +
+    currentdate.getMinutes() +
+    ":" +
+    currentdate.getSeconds();
+
+  return datetime;
+}
+function getID() {
+  return id++;
+}
+
+let id = 1;
+
+module.exports = {
+  getRandomWeather,
+  getRandomForecast,
+  getDatetime,
+  getID,
+};
